services and impl

+++ src/main/java/tn/esprit/spring/repositories/ModuleRepository.java
@@ -3,5 +3,8 @@ package tn.esprit.spring.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.entities.Module;
 
-public interface ModuleRepository extends JpaRepository<Module, Integer> {
+import java.util.List;
+
+public interface ModuleRepository extends JpaRepository<Module, Long> {
+
 }
